+++ docs/changelog.rst
@@ -29,6 +29,8 @@ Changed:
 * External commands started with ``!`` no longer run in a sub-shell. To run commands
   with a sub-shell use ``:spawn`` instead.
 * The selected path in the library is centered as in thumbnail mode if possible.
+* The library always focuses the child directory when entering the parent directory via
+  ``:scroll left``.
 
 Fixed:
 ^^^^^^